Check version between app-templates and project.
PROJECT_VERSION could be detected if project is located in AGL gerrit repo
or with personal tags when in personal repository
This version is checked against app-templates version which is
aligned on AGL tags.
A warning is displayed if app-templates version is outdated related to
project_version.

Also a "dirty" flag is appended to the project version if
git repo isn't in sync with its remote counterpart. Then you could
know if see an installed widget on a target that got that "-dirty"
flag appended to the version that this widget include code that could
not be merged in the git repository.

Bug-AGL: SPEC-951

Change-Id: I99687560d19d746c887fc7a095e71407bda34325
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>