start-on-target: uses RSYNC_PREFIX path for config
Bindings using the controller need to have a right base
search path for the config files, loaded plugins, and other extra
files (lua, for instance).
Therefore, in the development workflow here, where the binding
is uploaded to the target at a typically unsusual location,
the afb-daemon is given the RSYNC_PREFIX/etc as a base through the
CONFIG_CONTROL_PATH env variable.

Change-Id: Ieb39174918f0d51985e1d63b9f7c6509a05f284a
Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>