Warning if not using wgtpkg-pack to make a widget
Also use zip host command instead of embedded CMake zip command
which could not be unzipped on a board. Using the host zip command
it allows to build the wgt file as a zip archive but it can't be
installed using afm-util like a real widget file but could still be
unzipped manually.

Change-Id: I5e13c27164294d99de1918df8d84cb04fc439317
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>