Create a test widget
Create a test widget in addition of the normal one when you use
any build type except of RELEASE. This widget will only
provide the test files (configuration + test + fixtures ...).
It joins the test framework binding to your binding or app then
you only have to access to your defined test API and launch the
tests.

The entry point is a bash script which is in charge to launch all
tests it finds.

Change-Id: I8f4a670f17fd4e4319c53a861074fb5e10b63aad
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>