Change default compilation options.
This changes the default build type to RELEASE as it would
be built using Yocto. Debug build should be activated if needed
temporary as the other type: coverage and profiling.

Doing this you are able to detect compilation error that would happens
when building with Yocto and if you need to debug your program you only
need to recompile changing temporary the build type to DEBUG.

Bugs-AGL: SPEC-1644

Change-Id: I1823509a1a48415d3d440dd84d1e43ef48736ebe
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>